Move error handler middleware to a separate file
